# documentation/documentation/openapi.yaml
openapi: 3.0.0
info:
  title: Chat Application APIs
  description: A set of APIs for managing members, channels, and messages in a chat application.
  version: "1.0.0"
servers:
  - url: https://api.example.com/v1
    description: Production server
...
paths:
  /members:
    post:
      tags:
        - Members
      summary: Create a new member
      description: Create a new member in the `/members` table.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberCreateRequest'
      responses:
        '201':
          description: Member created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: Bad Request
  /members/{memberId}/deactivate:
    patch:
      tags:
        - Members
      summary: Deactivate a member
      description: Deactivates a member, changing their `isActive` state to `false` or removing any access privileges.
      parameters:
        - in: path
          name: memberId
          required: true
          schema:
            type: string
            format: uuid
          description: The GUID of the member
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Optional reason for deactivation
              example:
                reason: "Member requested account deactivation"
      responses:
        '200':
          description: Member deactivated successfully
        '404':
          description: Member not found
...
  /channels/{channelId}/members/{memberId}:
    delete:
      tags:
        - Channels
      summary: Remove a member from a channel
      description: Removes the specified member from the given channel, effectively deleting their `sessionMembers` record.
      parameters:
        - in: path
          name: channelId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: memberId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Member removed from the channel (no content returned)
        '404':
          description: Channel or Member not found
...
components:
  schemas:
...
    MemberCreateRequest:
      type: object
      properties:
        name:
          type: string
        nickname:
          type: string
        type:
          type: string
          description: Member type, e.g. admin, user, etc.
        tag:
          type: string
      required:
        - name
      example:
        name: John Doe
        nickname: johnny
        type: user
        tag: developer
    Member:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        nickname:
          type: string
        type:
          type: string
        tag:
          type: string
        createdAt:
          type: string
          format: date-time
      example:
        id: 123e4567-e89b-12d3-a456-426614174000
        name: John Doe
        nickname: johnny
        type: user
        tag: developer
        createdAt: 2024-01-01T00:00:00Z

    # ============== CHANNELS ==============
...